player controller, game state etc added, as well as prefabs and included in the scene
