+++ Assets/RTSGameMode.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8a0b324d541057346801ac35134e6974
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
