+++ Assets/RTSGameMode/RTSPlayerState.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8efb085658ce7e245aea1982adaa609e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
