+++ Assets/RTSGameMode/RTSSpectatorController.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fbbde7dbffad612448bf4ddcf19a36f4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
