+++ Assets/RTSGameMode/RTSSpectatorState.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8677abc3a7c098e48a105427fe9be247
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
