+++ Assets/RTSGameMode/RTSTeamState.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 97343f0d9372c4145afeebd1ec2fb6b7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
