+++ Assets/Scripts/Shared/Network/RTSPlayerState.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// Contains server data that is player specific
+public class RTSPlayerState : PlayerState
+{
+    public TeamState team;
+
+
+}
