+++ Assets/Scripts/Shared/Network/RTSPlayerState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29b0ea8fdcbd1bd4c8810e2974a2590b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
