+++ Assets/Scripts/Shared/Network/RTSSpectatorController.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class RTSSpectatorController : MonoBehaviour
+{
+    [HideInInspector]
+    public RTSSpectatorState playerState;
+
+
+}
