+++ Assets/Scripts/Shared/Network/RTSSpectatorController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3aec135f92c72304280b22901d29042b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
