+++ Assets/Scripts/Shared/Network/RTSSpectatorState.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class RTSSpectatorState : PlayerState
+{
+
+
+
+}
