+++ Assets/Scripts/Shared/Network/RTSSpectatorState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d5c60224a52d9614f8466132eccf6d4c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
